requirements.txt
annotated-types==0.7.0
anyio==4.10.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==41.0.8
fastapi==0.104.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
pycparser==2.22
pydantic==2.5.0
pydantic_core==2.33.2
PyMySQL==1.1.0
python-dotenv==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.23
sqlmodel==0.0.24
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn[standard]==0.24.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6